refactor(app) change folder name from api to modules
